refactor: add shared types workspace
