+++ apps/api/package.json
@@ -14,6 +14,7 @@
         "@opentelemetry/auto-instrumentations-node": "^0.77.0",
         "@opentelemetry/exporter-trace-otlp-http": "^0.219.0",
         "@opentelemetry/sdk-node": "^0.219.0",
+        "@sahidawa/types": "*",
         "@supabase/supabase-js": "^2.108.2",
         "bullmq": "^5.79.2",
         "compression": "^1.8.1",
