+++ apps/web/package.json
@@ -20,6 +20,7 @@
         "@opentelemetry/sdk-trace-web": "^2.8.0",
         "@radix-ui/react-tooltip": "^1.2.11",
         "@sahidawa/shared": "*",
+        "@sahidawa/types": "*",
         "@supabase/auth-helpers-nextjs": "^0.15.0",
         "@supabase/ssr": "^0.12.0",
         "@supabase/supabase-js": "^2.110.0",
