+++ packages/types/src/lasa.types.ts
@@ -1,0 +1,7 @@
+export type LasaMatchType = "sound-alike" | "look-alike";
+
+export interface LasaMatch {
+    name: string;
+    type: LasaMatchType;
+    score: number;
+}
